refactor: improve fcm routes structure
