+++ backend-go/internal/controller/user.go
@@ -3,26 +3,11 @@ package controller
 import (
 	"net/http"
 
+	db "github.com/LambdaIITH/Dashboard/backend/internal/db"
 	helpers "github.com/LambdaIITH/Dashboard/backend/internal/helpers"
+	schema "github.com/LambdaIITH/Dashboard/backend/internal/schema"
 	"github.com/gin-gonic/gin"
 )
-
-type UserStruct struct {
-	ID          int    `json:"id"`
-	Email       string `json:"email"`
-	Name        string `json:"name"`
-	Cr          string `json:"cr"`
-	PhoneNumber string `json:"phone_number"`
-}
-
-type UserUpdate struct {
-	PhoneNumber string `json:"phone_number"`
-}
-
-type FCMTokensUpdate struct {
-	Token      string `json:"token"`
-	DeviceType string `json:"device_type"`
-}
 
 func User(c *gin.Context) {
 	userId, err := helpers.GetUserID(c.Request)
@@ -30,7 +15,7 @@ func User(c *gin.Context) {
 		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
 	}
 
-	userDetails := helpers.GetUser(userId)
+	userDetails := db.GetUser(c, userId)
 	c.JSON(http.StatusOK, userDetails)
 
 }
@@ -41,13 +26,13 @@ func UpdateUser(c *gin.Context) {
 		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
 	}
 
-	var userUpdate UserUpdate
+	var userUpdate schema.UserUpdate
 	if err := c.ShouldBindJSON(&userUpdate); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
 
-	userDetails := helpers.UpdatePhone(userId, userUpdate.PhoneNumber)
+	userDetails := db.UpdatePhone(c, userId, userUpdate.PhoneNumber)
 	c.JSON(http.StatusOK, userDetails)
 }
 
@@ -57,13 +42,13 @@ func UpdateUserFCMToken(c *gin.Context) {
 		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
 	}
 
-	var fcmTokenUpdate FCMTokensUpdate
+	var fcmTokenUpdate schema.FCMTokensUpdate
 	if err := c.ShouldBindJSON(&fcmTokenUpdate); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
 
-	if helpers.UpsertFCMToken(userId, fcmTokenUpdate.Token, fcmTokenUpdate.DeviceType) {
+	if db.UpsertFCMToken(c, userId, fcmTokenUpdate.Token, fcmTokenUpdate.DeviceType) {
 		c.JSON(http.StatusOK, gin.H{"message": "FCM Token updated successfully"})
 	} else {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update FCM Token"})
